+++ src/sbuf.c
@@ -58,7 +58,7 @@ _sbuf__alloc_capacity(usize capacity)
     }
 }
 static inline Exception
-_sbuf__grow_buffer(sbuf_c* self, u32 length)
+_sbuf__grow_buffer(sbuf_c* self, usize length)
 {
     sbuf_head_s* head = (sbuf_head_s*)(*self - sizeof(sbuf_head_s));
 
@@ -68,7 +68,7 @@ _sbuf__grow_buffer(sbuf_c* self, u32 length)
         return Error.overflow;
     }
 
-    u32 new_capacity = _sbuf__alloc_capacity(length);
+    usize new_capacity = _sbuf__alloc_capacity(length);
     head = mem$realloc(head->allocator, head, new_capacity);
     if (unlikely(head == NULL)) {
         *self = NULL;
@@ -177,7 +177,7 @@ cex_sbuf_clear(sbuf_c* self)
 }
 
 /// Returns string length from its metadata
-static u32
+static usize
 cex_sbuf_len(sbuf_c* self)
 {
     uassert(self != NULL);
@@ -188,7 +188,7 @@ cex_sbuf_len(sbuf_c* self)
 
 
 /// Returns string capacity from its metadata
-static u32
+static usize
 cex_sbuf_capacity(sbuf_c* self)
 {
     uassert(self != NULL);
@@ -335,9 +335,9 @@ cex_sbuf_append(sbuf_c* self, char* s)
         return Error.argument;
     }
 
-    u32 length = head->length;
-    u32 capacity = head->capacity;
-    u32 slen = strlen(s);
+    usize length = head->length;
+    usize capacity = head->capacity;
+    usize slen = strlen(s);
 
     // Try resize
     if (length + slen > capacity - 1) {
